gui: Remove superfluous nearby devices datalist shown on input
Currently, the datalist is displayed in a separate pop-up that is shown
as soon as the user puts focus on the device ID input field when adding
a new device. This pop-up effectively covers other parts of the GUI,
such as the list of nearby devices, and also the device ID explanation
and possible warning messages. On top of that, the datalist includes the
exact same device IDs that are already listed under the nearby devices.
The device IDs in the datalist are also cut off with no option to scroll
horizontally, meaning that it is impossible to view them in full.

For the above reasons, remove the datalist completely as superfluous to
the already present and better list of nearby devices.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>