feat(db): never forget deleted files from database by default
Change the current 6-month period to "0" in order to disable forgetting
deleted files from the database. The user can still configure
the option via the command line, however it is disabled by default. This
prevents unexpected cases of files being "undeleted" when connecting
devices that have been disconnected for a long period of time.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>