feat(config): change default MaxFolderConcurrency to 50% of CPU cores
Currently, the default value of MaxFolderConcurrency equals the total
number of CPU cores. This allows processing of a large number of folders
at the some time, however it also has detrimental effect when it comes
to hashing individual folders.

For example, in a system with 100 folders and a 16-core CPU, Syncthing
can scan 16 folders at once, but each of those 16 folders is given only
a single CPU thread to hash the files. This isn't a problem when 16
folders actually need to be hashed, yet the same is true even if a
single folder is being scanned and hashed, as the default value of
MaxFolderConcurrency will still prevent more than 1 CPU threads from
being used for the hashing.

Therefore, reduce the default value of MaxFolderConcurrency to
half of the available CPU cores. This way, with the same 16-core CPU,
only 8 folders will be scanned and hashed at once, however each of them
will be allowed to use 2 threads of the CPU. Accordingly, when hashing
just one folder, it will be able to use 4 threads as well, instead of
just a single one.

Furthermore, the minimum value of MaxFolderConcurrency remains at 4,
which means that the number of folders allowed to be scanned and hashed
will still not be limited too much on CPUs with lower numbers of cores.

Signed-off-by: Tomasz Wilczyński <twilczynski@naver.com>